% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{int_breaks}
\alias{int_breaks}
\title{Integer axis-break generator with fixed step}
\usage{
int_breaks(step = 1)
}
\arguments{
\item{step}{Numeric. Step size between breaks. Default 1.}
}
\value{
A function \code{function(x)} that returns a numeric vector of
  break positions.
}
\description{
Returns a function that, given a numeric range, produces integer break
positions at a fixed interval. Useful as the \code{breaks} argument in
\pkg{ggplot2} scale functions.
}
\examples{
br <- int_breaks(step = 5)
br(c(0, 23))  # 0 5 10 15 20
}
