% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{int_breaks_pretty}
\alias{int_breaks_pretty}
\title{Pretty integer axis-break generator}
\usage{
int_breaks_pretty(n = 5)
}
\arguments{
\item{n}{Integer. Approximate number of breaks. Default 5.}
}
\value{
A function \code{function(x)} that returns a numeric vector of
  integer break positions.
}
\description{
Returns a function that uses \code{\link[base]{pretty}} to choose
approximately \code{n} integer break positions within a range. Useful as
the \code{breaks} argument in \pkg{ggplot2} scale functions.
}
\examples{
br <- int_breaks_pretty(n = 4)
br(c(0, 100))
}
